requirements.txt
pygame-ce>=2.4.0

